test: use dedicated context factories in tests
Currently the context factory used in tests creates an `AppContext`, but the middleware tests require different, simpler forms. The dataset middleware test doesn't actually test whether it works or not as datasets on the request and response already exist. This separates it into smaller factories, so the right context can be created for each test.